pyproject.toml
[tool.poetry]
name = "fastapi-from-php"
version = "0.1.0"
description = ""
authors = ["Artur Świerc <aswierc@gmail.com>"]
readme = "README.md"
packages = [{include = "python-ddd-architecture-for-php-devs"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
kink = "^0.6.5"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4"
psycopg2 = "^2.9"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
fastapi-events = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

